--- src/lib/Components/Artist/Articles/__tests__/Article-tests.tsx
@@ -1,20 +1,0 @@
-import "react-native"
-
-import { renderWithWrappers } from "lib/tests/renderWithWrappers"
-import React from "react"
-
-import Article from "../Article"
-
-it("renders without throwing an error", () => {
-  const article = {
-    thumbnail_title: "Something Happened",
-    href: "artsy.net/something-happened",
-    author: {
-      name: "John Berger",
-    },
-    thumbnail_image: {
-      url: "artsy.net/image-url",
-    },
-  }
-  renderWithWrappers(<Article article={article as any} />)
-})
